requirements.txt
aiofiles==23.1.0
anyio==3.7.1
asgiref==3.7.2
beautifulsoup4==4.12.2
certifi==2023.5.7
charset-normalizer==3.2.0
colorama==0.4.6
Django==4.2.3
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
nest-asyncio==1.5.6
orjson==3.9.2
Pillow==10.0.0
requests==2.31.0
sniffio==1.3.0
soupsieve==2.4.1
sqlparse==0.4.4
tqdm==4.65.0
twitter-api-client==0.10.11
tzdata==2023.3
urllib3==2.0.3
websockets==11.0.3
